Beefs up Further Reading section in README
Lots of great material went into creating this rather simple role, so
the URLs are as much for my reference as they are for the community.
Riemann is a powerful piece of software that will take some time to get
get used to, particularly if you don't have any Clojure experience.