Bumps required Ansible version to '2'
The 'apt' module got a lot smarter in Ansible v2, particularly when
using the 'deb' parameter. It's no longer necessary to inspect the
result of the apt task and conditionally report changed. That behavior
is still necessary under Ansible 1.9.x, so let's update the role to
require v2 and simplify the task logic.